requirements.txt
streamlit==1.18.1
openai==0.27.8
python-dotenv==1.0.0
PyMuPDF==1.23.1
python-docx==0.8.11